synapse/modules/container_app_job/docker/requirements.txt
azure-storage-blob==12.24.0
pymssql==2.2.12
pyarrow==17.0.0
